EP-22: Add the functionality of transferring the task to the status completed on the home page

Approved-by: Dzianis Buhara
Approved-by: Angelika Kozlova

+++ src/js/events/click-on-taskDeleteBtn.js
@@ -8,32 +8,32 @@ export default class ClickOnTaskDeleteBtn {
   }
 
   initEvent() {
-    const taskContainer = document.querySelector('.tasks-wrapper');
+    const taskContainer = document.querySelector('.action-wrapper');
     if (taskContainer) {
       taskContainer.addEventListener('click', (e) => {
         if (e.target.classList.contains('task__item-btn')) {
-          this.deleteTask(e);
+          if (e.target.classList.contains('btn-removed-task')) {
+            this.deleteTask(e);
+          }
         }
       });
     }
   }
 
   deleteTask(e) {
-    e.target.previousElementSibling.classList.toggle('removed');
-    this.updateStorage();
+    let currentTaskTitleHtml = e.target.parentNode.parentNode.querySelector('.task__item-text');
+    currentTaskTitleHtml.classList.toggle('removed');
+    this.updateStorage(currentTaskTitleHtml);
   }
 
-  updateStorage() {
+  updateStorage(currentTaskTitleHtml) {
     let savedTasks = JSON.parse(localStorage.getItem('_tasks')) ?? [];
-    const tasks = document.querySelectorAll('.task__item-text.removed');
     const newSavedTasks = [];
 
     savedTasks.forEach((item) => {
-      tasks.forEach((task) => {
-        if (task.innerHTML !== item.title) {
+        if (currentTaskTitleHtml.innerHTML !== item.title) {
           newSavedTasks.push(item);
         }
-      })
     })
 
     localStorage.setItem('_tasks', JSON.stringify(newSavedTasks));
